pyproject.toml
[build-system]
requires = ["poetry"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "fcov"
version = "0.1"
description = "Python functional coverage"
authors = []
license = "MIT"
readme = "README.adoc"

[tool.poetry.dependencies]
python = "^3.8"
rich = "^13.3.4"
sqlalchemy = "2.0.23"
gitpython = "3.1.41"
